+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   # Features
     * Added Get-NSRADIUSAuthenticationServer, New-NSRADIUSAuthenticationServer and Remove-NSRADIUSAuthenticationServer functions (via @iainbrighton)
     * Added Get-NSRADIUSAuthenticationPolicy, New-NSRADIUSAuthenticationPolicy and Remove-NSRADIUSAuthenticationPolicy functions (via @iainbrighton)
+    * Added -RADIUSAuthenticationPolicyName parameter to Add-NSVPNVirtualServerBinding function (via @iainbrighton)
+    * Added RADIUSAuthenticationPolicy -Binding parameter to Get-NSVPNVirtualServerBinding function (via @iainbrighton)
 
 ## 1.7.0 (2018-07-02)
   # Features